Move scanners outside workers pkg.
