+++ container_pipeline/scanners/base.py
